requirements.txt
timm==0.6.7
deepspeed==0.9.2
data
einops==0.6.1
ftfy==6.1.1
iopath==0.1.10
ipdb==0.13.13
numpy==1.24.3
peft==0.3.0
Pillow==9.5.0
PyYAML==6.0
regex==2022.10.31
pytorchvideo
fvcore
decord==0.6.0
tqdm==4.64.1
transformers==4.29.1
evaluate
openai-whisper
sentencepiece
librosa
# pip install --upgrade deepspeed